ai_agent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sentence-transformers==3.0.1
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.1